refactor: DescriptionField now implements 'sort by description'
