+++ src/Query/Filter/DescriptionField.ts
@@ -1,5 +1,6 @@
 import { getSettings } from '../../Config/Settings';
 import type { Task } from '../../Task';
+import type { Comparator } from '../Sorting';
 import { TextField } from './TextField';
 
 /**
@@ -26,5 +27,61 @@ export class DescriptionField extends TextField {
         // the global filter.
         const globalFilter = getSettings().globalFilter;
         return task.description.replace(globalFilter, '').trim();
+    }
+
+    public supportsSorting(): boolean {
+        return true;
+    }
+
+    /**
+     * Return a function to compare the description by how it is rendered in markdown.
+     *
+     * Does not use the MarkdownRenderer, but tries to match regexes instead
+     * in order to be simpler, faster, and not async.
+
+     */
+    public comparator(): Comparator {
+        return (a: Task, b: Task) => {
+            return DescriptionField.cleanDescription(a.description).localeCompare(
+                DescriptionField.cleanDescription(b.description),
+            );
+        };
+    }
+
+    /**
+     * Removes `*`, `=`, and `[` from the beginning of the description.
+     *
+     * Will remove them only if they are closing.
+     * Properly reads links [[like this|one]] (note pipe).
+     */
+    private static cleanDescription(description: string): string {
+        const globalFilter = getSettings().globalFilter;
+        description = description.replace(globalFilter, '').trim();
+
+        const startsWithLinkRegex = /^\[\[?([^\]]*)\]/;
+        const linkRegexMatch = description.match(startsWithLinkRegex);
+        if (linkRegexMatch !== null) {
+            const innerLinkText = linkRegexMatch[1];
+            // For a link, we have to check whether it has another visible name set.
+            // For example `[[this is the link|but this is actually shown]]`.
+            description =
+                innerLinkText.substring(innerLinkText.indexOf('|') + 1) + description.replace(startsWithLinkRegex, '');
+        }
+
+        const startsWithItalicOrBoldRegex = /^\*\*?([^*]*)\*/;
+        const italicBoldRegexMatch = description.match(startsWithItalicOrBoldRegex);
+        if (italicBoldRegexMatch !== null) {
+            const innerItalicBoldText = italicBoldRegexMatch[1];
+            description = innerItalicBoldText + description.replace(startsWithLinkRegex, '');
+        }
+
+        const startsWithHighlightRegex = /^==?([^=]*)==/;
+        const highlightRegexMatch = description.match(startsWithHighlightRegex);
+        if (highlightRegexMatch !== null) {
+            const innerHighlightsText = highlightRegexMatch[1];
+            description = innerHighlightsText + description.replace(startsWithHighlightRegex, '');
+        }
+
+        return description;
     }
 }
